update to device delete

+++ src/routes/api/devices/+server.ts
@@ -3,6 +3,7 @@ import { loadKey, loadSignedDeviceID } from "$lib/server/crypto";
 import { createKysely } from "$lib/server/db";
 import { error, json } from "@sveltejs/kit";
 import type { RequestHandler } from "./$types";
+import { sql } from "kysely";
 
 export const GET: RequestHandler = async (request) => {
   // get all devices linked to this account (requires cookie auth)
@@ -22,15 +23,12 @@ export const GET: RequestHandler = async (request) => {
         .where(
           "uid",
           "=",
-          db
-            .selectFrom("devicesToUsers")
-            .select("uid")
-            .where("did", "=", did)
+          db.selectFrom("devicesToUsers").select("uid").where("did", "=", did)
         )
     )
     .execute();
-  
-  return json(devices, {status: 200});
+
+  return json(devices, { status: 200 });
 };
 
 export const POST: RequestHandler = async ({
@@ -51,7 +49,8 @@ export const POST: RequestHandler = async ({
   if (isNaN(did)) throw error(400, "Invalid device id in query params");
 
   const { displayName, type, isOnline } = await request.json();
-  let updateObject: {displayName?: string, type?: string, isOnline?: number} = {};
+  let updateObject: { displayName?: string; type?: string; isOnline?: number } =
+    {};
   if (displayName) updateObject["displayName"] = displayName;
   if (type) updateObject["type"] = type;
   if (isOnline) updateObject["isOnline"] = isOnline;
@@ -95,10 +94,43 @@ export const DELETE: RequestHandler = async ({ platform, cookies, url }) => {
   if (isNaN(did)) throw error(400, "Invalid device id in query params");
 
   if (did == own_did) {
+    // get uid
+    const { uid } = await db
+      .selectFrom("devicesToUsers")
+      .select("uid")
+      .where("did", "=", own_did)
+      .executeTakeFirstOrThrow();
+
+    // delete deviceToUser mapping
+    const res1 = await db
+      .deleteFrom("devicesToUsers")
+      .where("did", "=", did)
+      .where("uid", "=", uid)
+      .executeTakeFirst();
+
+    console.log(`res1: ${JSON.stringify(res1)}`);
+    if (!res1) throw error(403, "You do not own this device");
+
     await db
       .deleteFrom("devices")
       .where("id", "=", did)
       .executeTakeFirstOrThrow();
+
+    // delete user if no devices left
+    const { count } = await db
+      .selectFrom("devicesToUsers")
+      .select(db.fn.countAll<number>().as("count"))
+      .where("uid", "=", uid)
+      .executeTakeFirstOrThrow();
+
+    console.log(`count: ${JSON.stringify(count)}`);
+    if (count == 0) {
+      console.log(`deleting user ${uid}`)
+      await db
+        .deleteFrom("users")
+        .where("id", "=", uid)
+        .executeTakeFirstOrThrow();
+    }
   } else {
     // delete deviceToUser mapping
     const res1 = await db
